Renamed all modules to correspond with naming conventions... I got it wrong
