+++ mapping/p99/ioc-tests/test_detector01.py
@@ -1,10 +1,9 @@
 import unittest
-from ca_common import *
+from cacommon import *
 
 DETECTOR_MODEL = "DC-152Q-C00-FI"
 PV_MODEL_NAME = "BL99P-EA-DET-01Model_RBV"
 PV_PORT_NAME = "BL99P-EA-DET-01PortName_RBV"
-
 
 def reset_cam_to_defaults():
     pass
